adding tokens to the most important forms and redirecting non-admin users if they write any admin page url

+++ src/Controller/admin/PostController.php
@@ -5,45 +5,53 @@ namespace App\Controller\admin;
 use App\Core\BaseController;
 use App\Entity\Post;
 use App\Repository\Manager\PostManager;
+use App\Services\AdminProtectionPart;
 use App\Services\PHPSession;
+use Ramsey\Uuid\Uuid;
 
 class PostController extends BaseController
 {
     
     /**
      * Form to create a new Post
-     *
-     * @return void
      */
     public function newPost()
     {
+        $adminProtectionPart = new AdminProtectionPart();
+        $adminProtectionPart->redirectNotAdmin();
+        $session = new PHPSession;
+        $uuid = Uuid::uuid4();
+        $uuid = $uuid->toString();
+        $session->set('token', $uuid);
         return $this->render('admin/newPost.html.twig', []);
     }
     
     /**
      * Add a new post to the database
-     *
-     * @return void
      */
-    public function addPost(string $title, string $heading, string $content, string $author)
+    public function addPost(string $title, string $heading, string $content, string $author, string $token)
     {
+        $adminProtectionPart = new AdminProtectionPart();
+        $adminProtectionPart->redirectNotAdmin();
         $fields = [$title, $heading, $content, $author];
+        $session = new PHPSession;
 
-        if($this->isSubmit('newPost') && $this->isValid($fields)) {
+        if($this->isSubmit('newPost') && $this->isValid($fields) && $token == $session->get('token')) {
+
+            $session->delete('token');
 
             $datePublication = $dateLastUpdate = date("Y-m-d H:i:s");
             $postManager = new PostManager('post');
 
             $post = new Post($title, $datePublication, $dateLastUpdate, $heading, $content, $author);
             $postManager->insert($post);
-            $session = new PHPSession;
+            
             $session->set('success', 'Votre nouveau post a bien été enregistré.');
 
             return $this->redirect('/admin-posts');
 
         } else {
 
-            $session = new PHPSession;
             $session->set('fail', 'Votre nouveau post n\'a pas été enregistré, une erreur dans le formulaire a été détectée.');
 
             return $this->redirect('/admin-posts');
@@ -53,51 +61,37 @@ class PostController extends BaseController
     
     /**
      * retrieves all published posts and forwards them to the page
-     *
-     * @return void
      */
     public function adminPosts()
     {
+        $adminProtectionPart = new AdminProtectionPart();
+        $adminProtectionPart->redirectNotAdmin();
         $adminPostManager = new PostManager('Post');
         $getAllPosts = $adminPostManager->getAll();
-
         $session = new PHPSession;
+        $uuid = Uuid::uuid4();
+        $uuid = $uuid->toString();
+        $session->set('token', $uuid);
 
-        if($session->get('success') != NULL)
-        {
-            $success = $session->get('success');
-            $session->delete('success');
-            return $this->render('admin/adminPosts.html.twig', [
-                "allPosts" => $getAllPosts,
-                'success' => $success
-            ]);
-
-        } elseif ($session->get('fail') != NULL)
-        {
-            $fail = $session->get('fail');
-            $session->delete('fail');
-            return $this->render('admin/adminPosts.html.twig', [
-                "allPosts" => $getAllPosts,
-                'fail' => $fail
-            ]);
-        } else
-        {
-            return $this->render('admin/adminPosts.html.twig', [
-                "allPosts" => $getAllPosts
-            ]);
-        }
+        return $this->render('admin/adminPosts.html.twig', [
+            "allPosts" => $getAllPosts
+        ]);
         
     }
     
     /**
      * retrieves the posts to modify and complete the field of the edit page
-     *
-     * @return void
      */
     public function editPost()
     {
+        $adminProtectionPart = new AdminProtectionPart();
+        $adminProtectionPart->redirectNotAdmin();
         $adminPostManager = new PostManager('Post');
         $getThisPost = $adminPostManager->getById($_GET['idPost']);
+        $session = new PHPSession;
+        $uuid = Uuid::uuid4();
+        $uuid = $uuid->toString();
+        $session->set('token', $uuid);
 
         return $this->render('admin/editPost.html.twig', [
             'getThisPost' => $getThisPost
@@ -112,15 +106,18 @@ class PostController extends BaseController
      * @param  string $heading
      * @param  string $content
      * @param  string $author
-     * @return void
+     * @param  string $token
      */
-    public function updatePost($id, string $title, string $heading, string $content, string $author)
+    public function updatePost($id, string $title, string $heading, string $content, string $author, string $token)
     {
-
+        $adminProtectionPart = new AdminProtectionPart();
+        $adminProtectionPart->redirectNotAdmin();
         $fields = [$title, $heading, $content, $author];
+        $session = new PHPSession;
 
-        if($this->isSubmit('editPost') && $this->isValid($fields)) {
+        if($this->isSubmit('editPost') && $this->isValid($fields) && $token == $session->get('token')) {
 
+            $session->delete('token');
             $dateLastUpdate = date("Y-m-d H:i:s");
             $postManager = new PostManager('post');
             $postData = $postManager->getById($id);
@@ -146,20 +143,31 @@ class PostController extends BaseController
      * Delete the post of the line
      *
      * @param  mixed $id
-     * @param  string $title
-     * @param  string $heading
-     * @param  string $content
-     * @param  string $author
-     * @return void
+     * @param  string $token
      */
-    public function deletePost($id)
+    public function deletePost($id = NULL, string $token = NULL)
     {
-        $deletePostManager = new PostManager('post');
-        echo $deletePostManager->delete($id);
-
+        if($id == NULL || $token == NULL)
+        {
+            $this->redirect('/erreur-403');
+        }
         $session = new PHPSession;
-        $session->set('success', 'Votre post a bien été supprimé.');
+        if($token == $session->get('token'))
+        {
+            $session->delete('token');
+            $adminProtectionPart = new AdminProtectionPart();
+            $adminProtectionPart->redirectNotAdmin();
+            $deletePostManager = new PostManager('post');
+            echo $deletePostManager->delete($id);
 
-        return $this->adminPosts();
+            $session = new PHPSession;
+            $session->set('success', 'Votre post a bien été supprimé.');
+
+            return $this->redirect('/admin-posts');
+        } else
+        {
+            return $this->redirect('/');
+        }
+        
     }
 }
